# internal/model/category.go
package model

type Category struct {
	ID            string     `json:"id" example:"1"`
	Name          string     `json:"name" example:"Laptop"`
	Slug          string     `json:"slug" example:"laptop"`
	ParentID      *string    `json:"-"`
	SubCategories []Category `json:"sub_categories,omitempty"`
}